moved Header/Footer components into logical folder, added main layout to keep footer at bottom, added some conditional rendering
