+++ src/App.js
@@ -1,19 +1,34 @@
 import './App.css';
-import Header from './components/header';
-import Footer from './components/footer';
+// import Header from './components/header';
+// import Footer from './components/MainLayout/Footer';
+import MainLayout from './components/MainLayout/Layout';
 
-function App() {
+// function App() {
+//   return (
+//     <div className="App">
+//       {/* <header className="App-header">
+//         <p>
+//           Welcome to Move Mentor! We're still working on this project. Stay tuned.
+//         </p>
+//       </header> */}
+//       <Header />
+//       <Footer />
+//     </div>
+//   );
+// }
+
+// export default App;
+
+const App = () => {
   return (
-    <div className="App">
-      {/* <header className="App-header">
-        <p>
-          Welcome to Move Mentor! We're still working on this project. Stay tuned.
-        </p>
-      </header> */}
-      <Header />
-      <Footer />
+    <div>
+      <MainLayout>
+        <p>Some Content Here</p>
+      </MainLayout>
+
     </div>
   );
-}
+};
 
 export default App;
+
