+++ src/components/MainLayout/header.css
@@ -1,0 +1,7 @@
+.header-teacher {
+  background-color: #dec4f3; /* Purple color for teachers (admins) */
+}
+
+.header-student {
+  background-color: #bcd5cf; /* Green color for students (non-admins) */
+}
