+++ src/components/MainLayout/header.jsx
@@ -1,18 +1,22 @@
 import Container from 'react-bootstrap/Container';
 import Navbar from 'react-bootstrap/Navbar';
-import HomeIcon from '../images/home_icon.png';
+import HomeIcon from '../../images/home_icon.png';
 import { Image } from 'react-bootstrap';
+import './header.css'
 
 
-function Header() {
+const Header = ({ userRole }) => {
+  const headerClass = userRole === 'teacher' ? 'header-teacher' : 'header-student';
+
   return (
-    <Navbar expand="lg" style={{backgroundColor: '#BCD5CF'}}>
+    <Navbar className={headerClass} expand="lg">
       <Container>
-        <a href="https://www.example.com" target="_blank" rel="noopener noreferrer">
+        <a href="#home">
           <Image src={HomeIcon} alt="Home icon to return home" fluid /></a>
       </Container>
     </Navbar>
   );
 }
+
 
 export default Header;
